# service/src/main/java/iuresti/training/peopleandcars/controller/CarsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package iuresti.training.peopleandcars.controller;

import iuresti.training.peopleandcars.modelapi.Car;
import iuresti.training.peopleandcars.modelapi.People;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-01-27T16:20:47.279-06:00[America/Chicago]")
@Validated
public interface CarsApi {

    /**
     * POST /cars : Add a new car
     * Create a car
     *
     * @param car Create a Car (required)
     * @return Successfull operation (status code 201)
     *         or Unexpected errror (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/cars",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml" }
    )
    ResponseEntity<Car> createCar(
         @Valid @RequestBody Car car
    );


    /**
     * DELETE /cars/{car_vin} : Delete a car by VIN
     * Delete a person by VIN
     *
     * @param carVin Car VIN (required)
     * @return Successfull operation (status code 204)
     *         or Car not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/cars/{car_vin}"
    )
    ResponseEntity<Void> deleteCar(
         @PathVariable("car_vin") String carVin
    );


    /**
     * GET /cars/{car_vin}/people : Find all people owning a car
     * Returns all people associate to a car
     *
     * @param carVin Car VIN (required)
     * @return Successfull operation (status code 200)
     *         or Car not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/cars/{car_vin}/people",
        produces = { "application/json", "application/xml" }
    )
    ResponseEntity<List<People>> getAllCarPeople(
         @PathVariable("car_vin") String carVin
    );


    /**
     * GET /cars : Get all cars
     * Returns all cars
     *
     * @return successful operation (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/cars",
        produces = { "application/json", "application/xml" }
    )
    ResponseEntity<List<Car>> getAllCars(
        
    );


    /**
     * GET /cars/{car_vin} : Find car by VIN
     * Return a single car
     *
     * @param carVin Car VIN (required)
     * @return Successfull operation (status code 200)
     *         or Car not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/cars/{car_vin}",
        produces = { "application/json", "application/xml" }
    )
    ResponseEntity<Car> getCarByVin(
         @PathVariable("car_vin") String carVin
    );


    /**
     * PUT /cars/{car_vin} : Update an existing car
     * Upate an existing car
     *
     * @param carVin Car VIN (required)
     * @param car Update a car (required)
     * @return Successfull operation (status code 200)
     *         or Car not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/cars/{car_vin}",
        produces = { "application/json", "application/xml" },
        consumes = { "application/json", "application/xml" }
    )
    ResponseEntity<Car> updateCar(
         @PathVariable("car_vin") String carVin,
         @Valid @RequestBody Car car
    );

}